refactor: controller 계층 제거 및 패키지 분리
