+++ src/main/java/blackjack/card/Denomination.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.card;
 
 import java.util.List;
