+++ src/main/java/blackjack/user/Participants.java
@@ -1,9 +1,7 @@
-package blackjack.domain.user;
+package blackjack.user;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Participants {
