+++ src/main/java/blackjack/view/InputView.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.user.PlayerName;
+import blackjack.user.PlayerName;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
