+++ src/test/java/blackjack/domain/user/DealerTest.java
@@ -3,11 +3,14 @@ package blackjack.domain.user;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import blackjack.domain.Card;
-import blackjack.domain.CardDeck;
-import blackjack.domain.Denomination;
-import blackjack.domain.GameResult;
-import blackjack.domain.Suit;
+import blackjack.card.Card;
+import blackjack.card.CardDeck;
+import blackjack.card.Denomination;
+import blackjack.game.GameResult;
+import blackjack.card.Suit;
+import blackjack.user.Dealer;
+import blackjack.user.Player;
+import blackjack.user.PlayerName;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
