+++ src/test/java/blackjack/domain/user/PlayerTest.java
@@ -2,10 +2,12 @@ package blackjack.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.Card;
-import blackjack.domain.CardDeck;
-import blackjack.domain.Denomination;
-import blackjack.domain.Suit;
+import blackjack.card.Card;
+import blackjack.card.CardDeck;
+import blackjack.card.Denomination;
+import blackjack.card.Suit;
+import blackjack.user.Player;
+import blackjack.user.PlayerName;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
